Moving settings test over, to preserve the app safety
